backend/pyproject.toml
[project]
name = "bank-diligence-platform"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic[email]",
    "pydantic-settings",
    "sqlalchemy",
    "psycopg[binary]",
    "alembic",
    "redis",
    "celery",
    "pyjwt",
    "python-multipart",
    "boto3",
    "pypdf",
    "reportlab",
    "pytesseract",
    "pillow",
    "pdf2image",
    "numpy",
    "opencv-python-headless",
    "pyyaml",
    "python-docx",
    "cryptography",
    "httpx",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

